+++ vue3-onsenui/src/components/VOnsCarousel.vue
@@ -1,7 +1,6 @@
 <template>
   <ons-carousel
-    :initial-index="index"
-    @postchange.self="$emit('update:index', $event.activeIndex)"
+    @postchange.self="$emit('update:activeIndex', $event.activeIndex)"
   >
     <div>
       <slot></slot>
@@ -12,27 +11,13 @@
 
 <script>
   import 'onsenui/esm/elements/ons-carousel.js';
-  import { hasOptions, deriveEvents } from '../mixins/index.js';
+  import { deriveEvents } from '../mixins/index.js';
 
   const name = 'v-ons-carousel';
 
   export default {
     name,
-    mixins: [hasOptions, deriveEvents(name)],
-    emits: ['update:index'],
-
-    props: {
-      index: {
-        type: Number
-      }
-    },
-
-    watch: {
-      index() {
-        if (this.index !== this.$el.getActiveIndex()) {
-          this.$el.setActiveIndex(this.index, this.options);
-        }
-      }
-    }
+    mixins: [deriveEvents(name)],
+    emits: ['update:activeIndex'],
   };
 </script>