+++ vue3-onsenui/src/components/VOnsFab.vue
@@ -5,7 +5,6 @@
 </template>
 
 <script>
-  /* This file was generated automatically by 'generate-components' task in bindings/vue/gulpfile.babel.js */
   import 'onsenui/esm/elements/ons-fab.js';
   import { deriveEvents, hidable, modifier } from '../mixins/index.js';