+++ vue3-onsenui/src/components/VOnsToolbar.vue
@@ -1,18 +1,20 @@
 <template>
-  <ons-toolbar :modifier="normalizedModifier">
+  <ons-toolbar :modifier="normalizedModifier" :inline="inline || null">
     <slot></slot>
   </ons-toolbar>
 </template>
 
 <script>
-  /* This file was generated automatically by 'generate-components' task in bindings/vue/gulpfile.babel.js */
   import 'onsenui/esm/elements/ons-toolbar.js';
-  import { deriveEvents, hidable, modifier } from '../mixins/index.js';
+  import { deriveEvents, modifier } from '../mixins/index.js';
 
   const name = 'v-ons-toolbar';
 
   export default {
     name,
-    mixins: [deriveEvents(name), hidable, modifier]
+    mixins: [deriveEvents(name), modifier],
+    props: {
+      inline: Boolean
+    }
   };
 </script>