+++ vue3-onsenui/CHANGELOG.md
@@ -8,6 +8,8 @@ dev
 
  * The ESM build is now the default when vue-onsenui is used with a bundler. When using the ESM build, components must be registered using `app.component` before they can be used.
  * For v-model, `modelValue` prop and `update:modelValue` event are now used instead of `modelProp` prop and `modelEvent` event. However, the `modelEvent` *prop* is unchanged.
+ * The `options.animation` and `options.animationOptions` props have been renamed to `animation` and `animationOptions` for all components.
+ * VOnsCarousel, VOnsTabbar, VOnsSegment: The `index` prop and `update:index` event have been renamed to `activeIndex` and `update:activeIndex`.
  * VOnsLazyRepeat: The renderItem prop should be a function that returns an *object* describing a Vue component.
 
  ### Bug Fixes